HPT-2440 checking for SCSI errors in get_alua_state method of UnixPathMixin
The get_alua_state method of UnixPathMixin could raise an error if the
ALUA state has changed (i.e an A/A replication has been created for
the parent device through InfiniBox and not through the host).
Catching those errors and suggesting the user to rescan (rescanning solves
those issues).